Added components, included backward links in pages, cleaned page structure, optimized api calls and fixed sequelize db structure
EmployeeLinks and productLinks components have been added and implemented to be reused in all the pages that need the preview of components and employees. Added api calls to include all the data that is needed to include the links to the other side of relationships in the content, fixed a bug with the foreign keys in the sequelize db structure.